requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
plotly>=5.3.0
jupyter>=1.0.0
openpyxl>=3.0.0
reportlab>=3.6.0
python-pptx>=0.6.212
